Fix: avoid import cycle
