+++ packages/liveblocks-react/src/liveblocks.tsx
@@ -34,12 +34,12 @@ import React, {
 import { useSyncExternalStore } from "use-sync-external-store/shim/index.js";
 import { useSyncExternalStoreWithSelector } from "use-sync-external-store/shim/with-selector.js";
 
+import { useIsInsideRoom } from "./contexts";
 import { byMostRecentlyUpdated } from "./lib/compare";
 import { makeThreadsFilter } from "./lib/querying";
 import { autoRetry, retryError } from "./lib/retry-error";
 import { useInitial, useInitialUnlessFunction } from "./lib/use-initial";
 import { use } from "./lib/use-polyfill";
-import { useIsInsideRoom } from "./room";
 import type {
   InboxNotificationsState,
   LiveblocksContextBundle,
